Create 'query_classifier.py' with the two query classifier classes from 'pipeline.py'.
